refactor: check functions properties' flags

+++ Dumper/UnrealTypes.h
@@ -98,7 +98,7 @@ public:
 
 	inline FString(const wchar_t* WChar)
 	{
-		MaxElements = NumElements = *WChar ? std::wcslen(WChar) + 1 : 0;
+		MaxElements = NumElements = *WChar ? static_cast<int32>(std::wcslen(WChar)) + 1 : 0;
 
 		if (NumElements)
 		{
